Change the way routers work
Use a new function to create a new router, then modify it in the plugin. This way plugins can use as many routers as they want, and they won't interfere with other plugins. Routers default paths are the plugin's name, but plugins can get around this when creating the router by explicitly setting the argument to something else, or to false for no path (for the web root).