Fix emotes
Emotes are entered in steam with standard colons at the ends, however steam servers send a character that looks similar, but isn't the same.
Since this isn't a character that users will ever actually be *using*, we'll just replace them all inside the trigger itself.
note to self: @efreak - search for these in logTrigger so we can display them in the live logger.