Change how git info is gotten
Use the git binary on the system; if it's not installed then we're obviously not in git.
Long string with git info is now available in chatBot.version.string
Short string without git info is now available in chatBot.version.short
Short string still replaces -dev with -git if in a repo.